+++ src/input/eventDetector/EventDetector.h
@@ -4,6 +4,13 @@
 #include "../../application/Randomizer.h"
 #include "../../application/structs/LocationData.h"
 
+enum LastStoryState
+{
+    LastStoryNotStarted,
+    LastStoryStarted,
+    LastStoryCompleted
+};
+
 
 class EventDetector
 {
@@ -11,6 +18,7 @@ public:
     explicit EventDetector(Randomizer& randomizer);
     void OnPlayingFrame() const;
     void OnLevelEmblem(int character, int level, int mission);
+    LastStoryState lastStoryState = LastStoryNotStarted;
 
 private:
     Randomizer& _randomizer;