+++ src/output/menuManager/MenuManager.cpp
@@ -3,11 +3,7 @@
 #include "../../pch.h"
 
 
-MenuManager::MenuManager()
-{
-    // FunctionHook<BOOL> blockTrialMenu(0x506780, []() -> BOOL { return false; });
-    FunctionHook<BOOL> blockMissionMenu(0x506410, []() -> BOOL { return false; });
-}
+
 
 void MenuManager::UnlockCharacterSelection(int eventAddress)
 {